added fpsViewer class, and FINALLY fixed bug with multiple led position.

+++ src/scene_objects/dgMultipleLedObject.h
@@ -42,7 +42,7 @@ public:
 	
 	int	numOfCols, numOfRows;
 	float xPadding, yPadding;
-	int xCentroid, yCentroid;
+	float xCentroid, yCentroid;
 	
 };
